+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 - added snyk security scans workflow
+- added managed mlflow to `sagemaker-studio` module
 
 ### **Changed**
 - enforce ssl on pipeline bucket in sagemaker templates module